+++ src/ksi/io.c
@@ -207,9 +207,9 @@ int KSI_RDR_isEOF(KSI_RDR *rdr) {
 	return rdr->eof;
 }
 
-static int readFromFile(KSI_RDR *rdr, unsigned char *buffer, const size_t size, unsigned int *readCount) {
+static int readFromFile(KSI_RDR *rdr, unsigned char *buffer, const size_t size, size_t *readCount) {
 	KSI_ERR err;
-	int count;
+	size_t count;
 
 	KSI_PRE(&err, rdr != NULL) goto cleanup;
 	KSI_PRE(&err, buffer != NULL) goto cleanup;
@@ -230,9 +230,9 @@ cleanup:
 	return KSI_RETURN(&err);
 }
 
-static int readFromMem(KSI_RDR *rdr, unsigned char *buffer, const size_t size, unsigned int *readCount) {
+static int readFromMem(KSI_RDR *rdr, unsigned char *buffer, const size_t size, size_t *readCount) {
 	KSI_ERR err;
-	unsigned int count;
+	size_t count;
 
 	KSI_PRE(&err, rdr != NULL) goto cleanup;
 	KSI_PRE(&err, buffer != NULL) goto cleanup;
@@ -261,7 +261,7 @@ cleanup:
 }
 
 
-int KSI_RDR_read_ex(KSI_RDR *rdr, unsigned char *buffer, const size_t bufferLength, unsigned int *readCount)  {
+int KSI_RDR_read_ex(KSI_RDR *rdr, unsigned char *buffer, const size_t bufferLength, size_t *readCount)  {
 	KSI_ERR err;
 	int res;
 
@@ -291,10 +291,10 @@ cleanup:
 	return KSI_RETURN(&err);
 }
 
-int KSI_RDR_read_ptr(KSI_RDR *rdr, unsigned char **ptr, const size_t len, unsigned int *readCount) {
+int KSI_RDR_read_ptr(KSI_RDR *rdr, unsigned char **ptr, const size_t len, size_t *readCount) {
 	KSI_ERR err;
 	unsigned char *p = NULL;
-	int count = 0;
+	size_t count = 0;
 
 	KSI_PRE(&err, rdr != NULL) goto cleanup;
 	KSI_PRE(&err, ptr != NULL) goto cleanup;
@@ -369,7 +369,7 @@ int KSI_RDR_verifyEnd(KSI_RDR *rdr) {
 	KSI_ERR err;
 	int res;
 	unsigned char *buf = NULL;
-	unsigned int buf_len = 0;
+	size_t buf_len = 0;
 
 	KSI_PRE(&err, rdr != NULL) goto cleanup;
 	KSI_BEGIN(rdr->ctx, &err);
